announcer/relayhostserv: Log user's current vhost
Added a function reporting user's current vhost on vhost request.

Added to make assigning vhosts easier and faster as opers will now be able to see when the person was assigned their vhost the last time and what their current vhost is.

Fixed the line endings to be the same as before.